Redesigned search UI and updated java file accordingly
I redesigned the search UI to make searching easier. I deleted everything but a single text box that when typed in, will update a recycler view with book entries in our Firebase database. I haven't finished the code quiet yet, but it will come together within the next few hours from the time of this commit.